Avoid Travis build failure
Travis CI doesn't have IPv6 connectivity to the Internet but glibc
always resolves both A and AAAA record. This resolver behavior cannot be
changed because glibc doesn't have any options on it.

Besides, this version H2O chooses a destination address from the
resolved addresses at random so there is a chance to choose unreachable
IPv6 address. It causes "failed to connect" error.

en.wikipedia.org which has both A and AAAA records in a test code just
causes the behavior above. To avoid the issue, this changes the address
to h2o.examp1e.net which only has A record