Avoid host resolving failure error on Travis CI
127.0.0.1.XIP.IO is used in a test code and it is basically resolvable
but it seems to be unstable and often fails. It may be caused by case
sensitiveness of the host name though the Internet domain name is case
insensitive but the reason is unclear.

Anyway, there is no way to avoid this annoying issue except adding it to
/etc/hosts now.

This adds a rule to append the host name to /etc/hosts on Travis CI
machine.